refactor: remove duplicate codes AFAP
add useTransactionHistory hook
