+++ packages/extension-polkagate/src/popup/history/partials/HistoryItem.tsx
@@ -1,8 +1,9 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
 /* eslint-disable react/jsx-max-props-per-line */
+
+import type { TransactionDetail } from '../../../util/types';
 
 import { ArrowForwardIosRounded as ArrowForwardIosRoundedIcon } from '@mui/icons-material';
 import { Container, Grid, IconButton, Typography } from '@mui/material';
@@ -10,7 +11,6 @@ import React, { useCallback, useMemo, useState } from 'react';
 
 import { FormatBalance2 } from '../../../components';
 import { useTranslation } from '../../../hooks';
-import type { TransactionDetail } from '../../../util/types';
 import { amountToMachine, toShortAddress, upperCaseFirstChar } from '../../../util/utils';
 import Detail from '../Detail';
 
@@ -25,7 +25,7 @@ interface Props {
   chainName: string | undefined;
 }
 
-export default function HistoryItem({ anotherDay, chainName, date, decimal, formatted, info, token }: Props): React.ReactElement {
+export default function HistoryItem ({ anotherDay, chainName, date, decimal, formatted, info, token }: Props): React.ReactElement {
   const { t } = useTranslation();
   const [showDetail, setShowDetail] = useState<boolean>(false);
 
@@ -41,24 +41,26 @@ export default function HistoryItem({ anotherDay, chainName, date, decimal, form
     }
 
     if (info?.from?.address === formatted) {
-      return `${t('To')}: ${info?.to?.name || toShortAddress(info?.to?.address as string)}`;
+      return `${t('To')}: ${info?.to?.name || toShortAddress(info?.to?.address)}`;
     }
 
     if (info?.to?.address === formatted) {
       return `${t('From')}: ${info?.from?.name || toShortAddress(info?.from?.address)}`;
     }
+
     return undefined;
   }, [formatted, info?.from?.address, info?.from?.name, info.subAction, info?.to, t]);
 
   return (
     <Container disableGutters sx={{ marginTop: `${anotherDay ? 20 : -0.8}px` }}>
-      {anotherDay && <Grid item sx={{ fontSize: '14px', fontWeight: 400 }}>
-        {date}
-      </Grid>
+      {anotherDay &&
+       <Grid container item justifyContent='center' sx={{ fontSize: '14px', fontWeight: 400 }}>
+         {date}
+       </Grid>
       }
-      <Grid alignItems='center' container direction='column' item justifyContent='space-between' sx={{ '> .historyItems:last-child': { border: 'none' }, bgcolor: 'background.paper', border: '1px solid', borderColor: 'secondary.light', borderRadius: '5px' }}>
+      <Grid alignItems='center' container direction='column' item justifyContent='space-between' sx={{ '> .historyItems:last-child': { border: 'none' }, bgcolor: 'background.paper', borderRadius: '5px' }}>
         <Grid className='historyItems' container item py='5px' sx={{ borderBottom: '1px solid', borderBottomColor: 'secondary.light' }}>
-          <Grid container item sx={{ fontSize: '22px', fontWeight: 300 }} px='10px' xs={11}>
+          <Grid container item px='10px' sx={{ fontSize: '22px', fontWeight: 300 }} xs={11}>
             <Grid container item justifyContent='space-between'>
               <Grid item>
                 {action}
@@ -73,19 +75,19 @@ export default function HistoryItem({ anotherDay, chainName, date, decimal, form
               </Grid>
             </Grid>
             <Grid container item justifyContent='space-between'>
-              <Grid item sx={{ whiteSpace: 'nowrap', width: 'fit-content', maxWidth: '65%' }}>
+              <Grid item sx={{ maxWidth: '65%', whiteSpace: 'nowrap', width: 'fit-content' }}>
                 <Typography fontSize='16px' fontWeight={200} sx={{ overflow: 'hidden', textOverflow: 'ellipsis' }}>
                   {subActionToFrom}
                 </Typography>
               </Grid>
               <Grid item>
-                <Typography fontSize='16px' fontWeight={400} color={info.success ? 'green' : 'red'}>
+                <Typography color={info.success ? 'green' : 'red'} fontSize='16px' fontWeight={400}>
                   {info.success ? t<string>('Completed') : t<string>('Failed')}
                 </Typography>
               </Grid>
             </Grid>
           </Grid>
-          <Grid alignItems='center' container item sx={{ borderLeft: '1px solid', borderLeftColor: 'secondary.light' }} xs={1}>
+          <Grid alignItems='center' container item sx={{ borderLeft: '1px solid', borderLeftColor: 'divider' }} xs={1}>
             <IconButton
               onClick={_goToDetail}
               sx={{ p: 0 }}
@@ -107,4 +109,4 @@ export default function HistoryItem({ anotherDay, chainName, date, decimal, form
       }
     </Container>
   );
-};
+}
